requirements.txt
bitarray >= 0.8.3
cryptography >= 2.3.1
Pillow >= 5.3.0
bcrypt >= 3.1.4